services/marketdata-rlc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
prometheus-client==0.20.0
aiokafka==0.10.0
asyncpg==0.29.0
numpy==1.26.4
onnxruntime==1.18.1
PyYAML==6.0.2
